Add a multipart/form-data file upload
We wanted to test the multipart upload in clj-http.
Since the API for uploading files in twitter is gated behind additional
user authentication and we don't want to implement it for the sake of
this PoC, I created a small rails application which receives "test.txt"
on `/multipart` and saves it to a file.
This commit adds `multipart-to-localhost` - a function that is supposed
to write to this application.